refactor: split CSS and improve BEM rules
This does not introduce anything new, it is a rewrite of some of the CSS.

The current version follows BEM to some extent, with many loopholes and badly
split styles. This commit fixes this, the styles are now following the BEM
methodology closely and simplifies the composition of our styles.